refactor(text): 👷 remove monolith file
